feat(ui): reshape Home into Read tab (Phase 3)
Rename the tab from Home to Read and reshape the screen as the
instrument the owner reads. Replaces the StatusCard (loud, card-shaped,
greeted "All vitals within normal range" even when nothing was
happening) with a quiet one-line alert strip that only renders when
there is something to surface. Drops the QuickSymptomCard from Home —
the Log tab is the entry surface for owner-reported state.

Reorder so vitals (the primary read) leads, with deeper lenses (Active
Substances, Body Levels) below. Baseline-building card moves to the
bottom and stays conditional.

QuickSymptomCard moves to LogScreen since it is the sole consumer now.
StatusCard composable is deleted.
